Major overhaul of the Adapters used in MenuActivity
1 - Used Butterknife to inject the Views

2 - Deimplemented OnItemClickListener from MenuActivity and created
class MenuDetailPopup instead.

The switch case in the onItemClick method with almost identical
operations says that we can abstract out the enum. MenuDetailPopup is
used to handle the clicks. This class takes an enum type parameter. In
its contructor it get a reference to the array of enum items of the type
parameter. In the OnItemClick method it selects an item from the array
and gets the name and description of the menu item to be displayed.

Doing it this way eliminates the need for the ArrayAdapters and String
array that were defined as members of MenuActivity.

3 - Created interface MenuItem as a common interface over all three
enums defined in MenuActivity. This helps MenuDetailPopup to abstract
away which enum class it has.

4 - Created class EnumAdapter. This adapter is parametrized over any
enum, and it uses the method Class::getEnumConstants to get an array of
all instances of that enum, which it then passes to the ArrayAdapter
constructor.

5 - Created interface ViewDecorator to help the EnumAdapter be more
flexible. This eliminates the need of CustomFontArrayAdapter. Using
decorators is a lot more flexible, because we can compose as many
modification as we want, even dynamically, whereas creating a new
subclass for each style of Adapter will lead to a lot of classes.

6 - Changed the paddings used to 16dp. This is the recomended default
padding in most layouts. Notice that the left edge of the description
text lines up with the Dialog's title text.

7 - Used the proper attributes in menu_list_item.